% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitochondria_check.R
\name{add_mt}
\alias{add_mt}
\title{Add mitochondrial information to a Seurat object}
\usage{
add_mt(seuratobject)
}
\arguments{
\item{seuratobject}{single cell data in seurat format}
}
\description{
Add mitochondrial information to a Seurat object
}
