% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{add_multiple_mt}
\alias{add_multiple_mt}
\title{Add mitochondrial percentage to multiple seurat objects}
\usage{
add_multiple_mt(seuratlist)
}
\arguments{
\item{seuratlist}{list of seurat objects}
}
\value{
seurat object with mitochondrial percentages added to it
}
\description{
Add mitochondrial percentage to multiple seurat objects
}
