% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering_seurat_list}
\alias{clustering_seurat_list}
\title{Apply single cell clustering to an entire list of seurat objects}
\usage{
clustering_seurat_list(
  seurat_list,
  features_to_use,
  variable.selection.method,
  variable.selection.nfeatures,
  number_of_dimensions,
  resolution,
  tsne,
  umap
)
}
\arguments{
\item{seurat_list}{list of seurat objects}

\item{features_to_use}{all or variable}

\item{variable.selection.method}{variable selection method}

\item{variable.selection.nfeatures}{number of features to use in variable selection}

\item{number_of_dimensions}{number of dimensions to use for clustering}

\item{resolution}{clustering resolution}

\item{tsne}{true or false for tsne reduction}

\item{umap}{true or false for uman}
}
\description{
Apply single cell clustering to an entire list of seurat objects
}
