% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{investigate_cluster_number}
\alias{investigate_cluster_number}
\title{Analyze data for the appropriate number of clusters}
\usage{
investigate_cluster_number(
  seuratobject,
  variable.selection.method,
  variable.selection.nfeatures,
  max.number.clusters.to.check,
  jackstraw.approach
)
}
\arguments{
\item{seuratobject}{your normalized, transformed, etc... seurat object}

\item{variable.selection.method}{to select your highly variable features}

\item{variable.selection.nfeatures}{how many variable features you want to pull out}

\item{max.number.clusters.to.check}{the maximum number of components/clusters you want to look at}

\item{jackstraw.approach}{true or false to run a jackstraw analysis (takes some time)}
}
\value{
a list with information from both a principal component and jackstraw approach
}
\description{
Analyze data for the appropriate number of clusters
}
