% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_and_scale_sc_counts.R
\name{normalize_sc}
\alias{normalize_sc}
\title{Function to be called to normalize single cell data}
\usage{
normalize_sc(seuratobject, method, scale)
}
\arguments{
\item{seuratobject}{the seurat object you want normalized}

\item{method}{whichever normalization method you want to implement like "LogNormalize"}

\item{scale}{scale}
}
\value{
a seurat object with normalized count information
}
\description{
Function to be called to normalize single cell data
}
