% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_and_scale_sc_counts.R
\name{normalize_seurat_list}
\alias{normalize_seurat_list}
\title{Normalize an entire seurat list}
\usage{
normalize_seurat_list(seuratlist, method, scale)
}
\arguments{
\item{seuratlist}{your list of seurat objects you want normalized}

\item{method}{the normalization method of your choice}

\item{scale}{scale}
}
\value{
a list of seurat objects with normalized counts
}
\description{
Normalize an entire seurat list
}
