% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitochondria_check.R
\name{onesample_singlecell_QC_filtering}
\alias{onesample_singlecell_QC_filtering}
\title{Filter a single cell dataset based on molecule count, feature count, and mitochondrial percentage.}
\usage{
onesample_singlecell_QC_filtering(
  seuratobject,
  minimum_counts,
  minimum_features,
  max_mitochondrialperc
)
}
\arguments{
\item{seuratobject}{A single cell seurat object}

\item{minimum_counts}{The minimum number of molecules you want per cell}

\item{minimum_features}{The minimum number of genes you want per cell}

\item{max_mitochondrialperc}{The maximum mitochondrial percentage}
}
\value{
filtered seurat object
}
\description{
Filter a single cell dataset based on molecule count, feature count, and mitochondrial percentage.
}
