% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{qc_filtering}
\alias{qc_filtering}
\title{Filter one single cell dataset based on molecule count, feature count, and mitochondrial percentage.}
\usage{
qc_filtering(
  seuratobject,
  minimum_counts,
  minimum_features,
  max_mitochondrialperc
)
}
\arguments{
\item{seuratobject}{the seurat data}

\item{minimum_counts}{The minimum number of molecules you want per cell}

\item{minimum_features}{The minimum number of genes you want per cell}

\item{max_mitochondrialperc}{The maximum mitochondrial percentage}
}
\value{
filtered seurat object
}
\description{
Filter one single cell dataset based on molecule count, feature count, and mitochondrial percentage.
}
