% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin_singlecell_h5.R
\name{readin_singlecell_h5}
\alias{readin_singlecell_h5}
\title{Import a single cell h5 file}
\usage{
readin_singlecell_h5(filename, sampleid)
}
\arguments{
\item{filename}{Your h5 file and any needed file path}

\item{sampleid}{What you want the file to be called upon import}
}
\value{
A seurat object
}
\description{
Import a single cell h5 file
}
