% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{sctype_annotation}
\alias{sctype_annotation}
\title{Annotate the clusters for a seurat object}
\usage{
sctype_annotation(seuratobject, exhaustion_set, reduction_type, plot_title)
}
\arguments{
\item{seuratobject}{the object to annotate}

\item{exhaustion_set}{whether to include the exhaustion signatures}

\item{reduction_type}{tsne or umap}

\item{plot_title}{plot title}
}
\value{
plot with clusters defined
}
\description{
Annotate the clusters for a seurat object
}
