% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitochondria_check.R
\name{singlecell_QC_filtering}
\alias{singlecell_QC_filtering}
\title{Filter single cell data based on molecule count, feature count, and mitochondrial percentage.}
\usage{
singlecell_QC_filtering(
  seuratlist,
  minimum_counts,
  minimum_features,
  max_mitochondrialperc
)
}
\arguments{
\item{seuratlist}{A list of seurat objects}

\item{minimum_counts}{The minimum number of molecules you want per cell}

\item{minimum_features}{The minimum number of genes you want per cell}

\item{max_mitochondrialperc}{The maximum mitochondrial percentage}
}
\value{
filtered seurat object
}
\description{
Filter single cell data based on molecule count, feature count, and mitochondrial percentage.
}
