% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{features_counts_mitoperc_dotplot}
\alias{features_counts_mitoperc_dotplot}
\title{Plot the number of molecules, genes, and mitochondrial percentage per seurat object}
\usage{
features_counts_mitoperc_dotplot(
  seuratobject,
  min_feature_limit,
  min_mol_limit,
  plot_title
)
}
\arguments{
\item{seuratobject}{from seurat}

\item{min_feature_limit}{the minimum number of genes you want detected in each cell for QC purposes}

\item{min_mol_limit}{the minimum number of molecules you want detected in each cell for QC purposes}

\item{plot_title}{whatever you want the plot to be called}
}
\value{
a plot
}
\description{
Plot the number of molecules, genes, and mitochondrial percentage per seurat object
}
